renamed and moved GA_utils.py
